myproject/stockmonitor/requirements.txt
Django
mysqlclient
requests
pandas
Django>=3.0,<4.0
requests
pandas
mysqlclient
redis>=3.5
gunicorn>=20.0.4
django-extensions
flower
yfinance
django-cors-headers
prometheus_client
django-prometheus
python-decouple==3.8
djangorestframework
djangorestframework-simplejwt